+++ tomography_preprocessing/utils/file.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 from typing import List
 
@@ -10,3 +11,7 @@ def basename(path: Path):
 
 def glob(pattern: str) -> List[Path]:
     return list(Path().glob(pattern))
+
+
+def write_empty_file(filename: os.PathLike) -> None:
+    open(filename, mode='w').close()
